Fix getting permission data
The `mData` field has been removed from the `AppPermissionAdapter` class
in versions 5.10.x+ and up of the Play Store.

In this commit a hook was added to the constructor of the
`AppPermissionAdapter` class to create the `mData` object and set it
as an additional field if it failed to access it from the object.

The functions that access `mData` were then patched to look on the
native object first, then fall back to looking in the additional fields.

This allows the module to be compatible with v5.4.x to v6.0.5 (current)
of the Play Store.